feat(AutoF5): custom checks
