+++ src/main/kotlin/net/ccbluex/liquidbounce/features/module/modules/player/cheststealer/ModuleChestStealer.kt
@@ -21,8 +21,6 @@
 package net.ccbluex.liquidbounce.features.module.modules.player.cheststealer
 
 import net.ccbluex.liquidbounce.config.types.NamedChoice
-import net.ccbluex.liquidbounce.config.types.ValueType
-import net.ccbluex.liquidbounce.config.types.nesting.ToggleableConfigurable
 import net.ccbluex.liquidbounce.event.events.ScheduleInventoryActionEvent
 import net.ccbluex.liquidbounce.event.handler
 import net.ccbluex.liquidbounce.features.module.Category
@@ -30,15 +28,11 @@ import net.ccbluex.liquidbounce.features.module.ClientModule
 import net.ccbluex.liquidbounce.features.module.modules.player.cheststealer.features.FeatureChestAura
 import net.ccbluex.liquidbounce.features.module.modules.player.cheststealer.features.FeatureSilentScreen
 import net.ccbluex.liquidbounce.features.module.modules.player.invcleaner.*
-import net.ccbluex.liquidbounce.utils.collection.Filter
 import net.ccbluex.liquidbounce.utils.inventory.*
 import net.ccbluex.liquidbounce.utils.item.isMergeable
 import net.minecraft.client.gui.screen.Screen
 import net.minecraft.client.gui.screen.ingame.HandledScreen
 import net.minecraft.item.ItemStack
-import net.minecraft.screen.ScreenHandlerType
-import net.minecraft.text.Text
-import java.util.EnumSet
 import kotlin.math.ceil
 
 /**
@@ -64,71 +58,8 @@ object ModuleChestStealer : ClientModule("ChestStealer", Category.PLAYER) {
 //        PUT_BACK("PutBack"), TODO: Fix this
     }
 
-    private object CheckScreenHandlerType : ToggleableConfigurable(this, "CheckScreenHandlerType", enabled = true) {
-        private val types by registryList(
-            "Types",
-            hashSetOf(
-                ScreenHandlerType.GENERIC_9X3, ScreenHandlerType.GENERIC_9X6, ScreenHandlerType.SHULKER_BOX,
-            ),
-            ValueType.SCREEN_HANDLER
-        )
-        private val filter by enumChoice("Filter", Filter.WHITELIST)
-
-        fun canSteal(screen: HandledScreen<*>): Boolean {
-            return !enabled || filter(runCatching { screen.screenHandler.type }.getOrNull(), types)
-        }
-    }
-
-    init {
-        tree(CheckScreenHandlerType)
-    }
-
-    private object CheckTitle : ToggleableConfigurable(this, "CheckTitle", enabled = true) {
-        private val titles by multiEnumChoice(
-            "Titles",
-            EnumSet.of(
-                ContainerTitle.CHEST, ContainerTitle.LARGE_CHEST,
-                ContainerTitle.SHULKER_BOX, ContainerTitle.BARREL,
-            ),
-        )
-        private val filter by enumChoice("Filter", Filter.WHITELIST)
-
-        fun canSteal(screen: Screen): Boolean {
-            if (!enabled) return true
-
-            val titleString = screen.title.string
-
-            return when (filter) {
-                Filter.WHITELIST -> titles.any {
-                    Text.translatable(it.translatableKey).string == titleString
-                }
-                Filter.BLACKLIST -> titles.none {
-                    Text.translatable(it.translatableKey).string == titleString
-                }
-            }
-        }
-    }
-
-    init {
-        tree(CheckTitle)
-    }
-
-    @Suppress("unused")
-    private enum class ContainerTitle(override val choiceName: String, val translatableKey: String) : NamedChoice {
-        BARREL("Barrel", "container.barrel"),
-        BEACON("Beacon", "container.beacon"),
-        BLAST_FURNACE("BlastFurnace", "container.blast_furnace"),
-        BREWING_STAND("BrewingStand", "container.brewing"),
-        CHEST("Chest", "container.chest"),
-        LARGE_CHEST("LargeChest", "container.chestDouble"),
-        DISPENSER("Dispenser", "container.dispenser"),
-        DROPPER("Dropper", "container.dropper"),
-        ENDER_CHEST("EnderChest", "container.enderchest"),
-        FURNACE("Furnace", "container.furnace"),
-        HOPPER("Hopper", "container.hopper"),
-        SHULKER_BOX("ShulkerBox", "container.shulkerBox"),
-        SMOKER("Smoker", "container.smoker"),
-    }
+    private val checkScreenHandlerType = tree(CheckScreenHandlerTypeConfigurable(this))
+    private val checkScreenTitle = tree(CheckScreenTitleConfigurable(this))
 
     init {
         tree(FeatureChestAura)
@@ -377,7 +308,8 @@ object ModuleChestStealer : ClientModule("ChestStealer", Category.PLAYER) {
     }
 
     fun Screen.canBeStolen(): Boolean {
-        return running && this is HandledScreen<*> && CheckScreenHandlerType.canSteal(this) && CheckTitle.canSteal(this)
+        return running && this is HandledScreen<*> &&
+            checkScreenHandlerType.isValid(this) && checkScreenTitle.isValid(this)
     }
 
     private enum class ItemMoveMode(override val choiceName: String) : NamedChoice {
